feat: improve charts performance
