+++ src/pages/Home/components/index.ts
@@ -1,0 +1,3 @@
+export { default as LineChart } from './LineChart'
+export { default as BarChart } from './BarChart'
+export { default as HideSmall } from './HideSmall'
